apiserver/requirements.txt
pillow
cmake
wheel
scipy
scikit-image
dlib
opencv-python
